% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgcvn.R
\name{getgcvn}
\alias{getgcvn}
\title{Get local land cover}
\usage{
getgcvn()
}
\description{
Retrieves a raster file containing the land cover of Vietnam.
}
\examples{
library(raster)
landcover <- getgcvn()
plot(landcover)
}
