% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlandcover.R
\name{getlandcover}
\alias{getlandcover}
\title{Proportions of provinces per landcover}
\usage{
getlandcover(date = "2015-01-01", merge_hanoi = FALSE)
}
\arguments{
\item{date}{either text in the "YYYY-MM-DD" format (for example "2015-01-17"
for the 17th of January 2017), or a numeric format of the year (for example
2015). By default "2015-01-01".}

\item{merge_hanoi}{boolean indicating whether the provinces of Ha Noi and
Ha Son Binh (or Ha Noi and Ha Tay) should be merged.}
}
\value{
An object of class "SpatialPolygonsDataFrame".
}
\description{
Generates a provinces map of Vietnam for a given date with proportions of
each land cover category as attributes.
}
\details{
These maps are prepared by
}
\note{
argument \code{merge_hanoi} makes a difference only for \code{date}
before 2008-01-01.
}
\author{
Marc Choisy
}
