% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globcoverVN.R
\docType{package}
\name{globcoverVN}
\alias{globcoverVN}
\alias{globcoverVN-package}
\title{globcoverVN: land use of Vietnam from the GlobCover project.}
\description{
The globcover package provides land use data for Vietnam from the GlobCover
project. The data contain 20 types of land use.
}
