% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_legend.R
\docType{methods}
\name{show_legend}
\alias{show_legend}
\alias{show_legend,RasterLayer-method}
\title{Show legend}
\usage{
show_legend(object)

\S4method{show_legend}{RasterLayer}(object)
}
\arguments{
\item{object}{an object of type raster}
}
\description{
Print to screen the legend of a spatial object.
}
\examples{
library(raster)
landcover <- getgcvn()
leg <- show_legend(landcover)

colors <- leg$color
plot(landcover, col = colors, legend = FALSE, axes = FALSE)
l <- length(colors)
x <- ceiling(l / 2)
sel1 <- 1:x
sel2 <- (x + 1):l
legend("left", legend = leg$code[sel1], fill = colors[sel1], bty = "n")
legend("right", legend = leg$code[sel2], fill = colors[sel2], bty = "n")
leg
}
