% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globcoverVN.R
\docType{package}
\name{globcoverVN}
\alias{globcoverVN}
\alias{globcoverVN-package}
\title{globcoverVN: land cover of Vietnam from the GlobCover project.}
\description{
The globcover package provides land cover data for Vietnam from the GlobCover
project. The data contain 20 types of land cover.
}
