% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_legend.R
\docType{methods}
\name{show_legend}
\alias{show_legend}
\alias{show_legend,RasterLayer-method}
\title{Show legend}
\usage{
show_legend(object)

\S4method{show_legend}{RasterLayer}(object)
}
\description{
Print to screen the legend of a spatial object.
}

